+++ openstack/blockstorage/v3/snapshots/results.go
@@ -5,61 +5,29 @@ import (
 	"time"
 
 	"github.com/opentelekomcloud/gophertelekomcloud"
-	"github.com/opentelekomcloud/gophertelekomcloud/internal/extract"
-	"github.com/opentelekomcloud/gophertelekomcloud/pagination"
 )
 
-// Snapshot contains all the information associated with a Cinder Snapshot.
 type Snapshot struct {
 	// Unique identifier.
 	ID string `json:"id"`
-
 	// Date created.
 	CreatedAt time.Time `json:"-"`
-
 	// Date updated.
 	UpdatedAt time.Time `json:"-"`
-
 	// Display name.
 	Name string `json:"name"`
-
 	// Display description.
 	Description string `json:"description"`
-
 	// ID of the Volume from which this Snapshot was created.
 	VolumeID string `json:"volume_id"`
-
 	// Currect status of the Snapshot.
 	Status string `json:"status"`
-
 	// Size of the Snapshot, in GB.
 	Size int `json:"size"`
-
 	// User-defined key-value pairs.
 	Metadata map[string]string `json:"metadata"`
 }
 
-// CreateResult contains the response body and error from a Create request.
-type CreateResult struct {
-	commonResult
-}
-
-// GetResult contains the response body and error from a Get request.
-type GetResult struct {
-	commonResult
-}
-
-// DeleteResult contains the response body and error from a Delete request.
-type DeleteResult struct {
-	golangsdk.ErrResult
-}
-
-// SnapshotPage is a pagination.Pager that is returned from a call to the List function.
-type SnapshotPage struct {
-	pagination.LinkedPageBase
-}
-
-// UnmarshalJSON converts our JSON API response into our snapshot struct
 func (r *Snapshot) UnmarshalJSON(b []byte) error {
 	type tmp Snapshot
 	var s struct {
@@ -77,57 +45,4 @@ func (r *Snapshot) UnmarshalJSON(b []byte) error {
 	r.UpdatedAt = time.Time(s.UpdatedAt)
 
 	return err
-}
-
-// IsEmpty returns true if a SnapshotPage contains no Snapshots.
-func (r SnapshotPage) IsEmpty() (bool, error) {
-	volumes, err := ExtractSnapshots(r)
-	return len(volumes) == 0, err
-}
-
-func (r SnapshotPage) NextPageURL() (string, error) {
-	var s struct {
-		Links []golangsdk.Link `json:"snapshots_links"`
-	}
-	err := r.ExtractInto(&s)
-	if err != nil {
-		return "", err
-	}
-	return golangsdk.ExtractNextURL(s.Links)
-}
-
-// ExtractSnapshots extracts and returns Snapshots. It is used while iterating over a snapshots.List call.
-func ExtractSnapshots(r pagination.Page) ([]Snapshot, error) {
-	var res struct {
-		Snapshots []Snapshot `json:"snapshots"`
-	}
-	err := (r.(SnapshotPage)).ExtractInto(&res)
-	return res.Snapshots, err
-}
-
-// UpdateMetadataResult contains the response body and error from an UpdateMetadata request.
-type UpdateMetadataResult struct {
-	commonResult
-}
-
-// ExtractMetadata returns the metadata from a response from snapshots.UpdateMetadata.
-func (r UpdateMetadataResult) ExtractMetadata() (map[string]interface{}, error) {
-	if r.Err != nil {
-		return nil, r.Err
-	}
-	m := r.Body.(map[string]interface{})["metadata"]
-	return m.(map[string]interface{}), nil
-}
-
-type commonResult struct {
-	golangsdk.Result
-}
-
-// Extract will get the Snapshot object out of the commonResult object.
-func (r commonResult) Extract() (*Snapshot, error) {
-	var res struct {
-		Snapshot *Snapshot `json:"snapshot"`
-	}
-	err = extract.Into(raw.Body, &res)
-	return res.Snapshot, err
 }
